ze_brands/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
black==23.1.0
boto3==1.26.106
botocore==1.29.106
Cerberus==1.3.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.3
Django==4.1.7
django-graphql-jwt==0.3.4
django-redis==5.2.0
djangorestframework==3.14.0
exceptiongroup==1.1.1
Faker==12.0.1
fastdiff==0.3.0
flake8==6.0.0
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mccabe==0.7.0
mixer==7.2.2
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.0
pathspec==0.11.1
platformdirs==3.1.1
pluggy==1.0.0
promise==2.3
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.6.0
pytest==7.2.2
pytest-cov==4.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
requests==2.28.2
s3transfer==0.6.0
simplejson==3.19.1
six==1.16.0
snapshottest==0.6.0
sqlparse==0.4.3
termcolor==2.2.0
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
wasmer==1.1.0
wasmer-compiler-cranelift==1.1.0
